Add the AlphaTabView with its play button
This activity displays dynamically the notes of a passed MusicXML file

This activity has been written using views, not composables. The library
that I'm using heavily relies on side-effects using Event Emitters and
so I wasn't able to find a way to use composables here.

In the future if I grew more accustomed to JetPack Compose, I might
decide to rewriting this part - or the part of the library I'm using